+++ config/routes.rb
@@ -11,7 +11,9 @@ Rails.application.routes.draw do
     namespace :cp do
       root "dashboard#show"
       resources :content_types do
-        resources :entries
+        resources :entries do
+          patch :image_field, on: :member
+        end
       end
       resources :assets, except: [ :show ]
       resources :globals
@@ -22,6 +24,7 @@ Rails.application.routes.draw do
         resources :form_submissions, only: [ :show, :destroy ]
       end
       resource :site_settings, only: [ :show, :edit, :update ]
+      patch "site_settings/image_field", to: "site_settings#image_field", as: :site_settings_image_field
       resources :themes, only: [ :index, :create, :update ]
       resources :taxonomies do
         resources :terms, except: [ :index, :show ]